[CBR-377] Separate Plugins into 'Plugins' and 'LegacyPlugins'
This will make it easier to remove the legacy stuff once we move
forward. Also, this commits adds the missing and (required) plugins to
the new wallet server. A few implementations are still missing (update
notifier and acid state snapshots) but will come.

Note that the new kernel registers itself to the node and receives new
block directly, there's no more need for a syncing pluging. Similarly,
the kernel takes care of managing restoration of wallets, even after a
node restart. Regarding the re-submission, this is currently done in
activeLayer bracket and the submissionLayer. This could be moved in a
later stage to a plugin for more consistency
