+++ wallet-new/cardano-sl-wallet-new.cabal
@@ -128,6 +128,7 @@ library
                        Cardano.Wallet.Server
                        Cardano.Wallet.Server.CLI
                        Cardano.Wallet.Server.Plugins
+                       Cardano.Wallet.Server.LegacyPlugins
                        Cardano.Wallet.TypeLits
                        Cardano.Wallet.Types.UtxoStatistics
                        Cardano.Wallet.Client
