+++ wallet-new/src/Cardano/Wallet/Server.hs
@@ -24,22 +24,18 @@ import           Cardano.Wallet.WalletLayer (ActiveWalletLayer (..))
 -- Cardano monad because they just interfact with the Wallet object.
 walletServer :: ActiveWalletLayer IO
              -> RunMode
-             -> Server WalletAPI
+             -> Server NewWalletAPI
 walletServer w _ =
-         v0Handler
-    :<|> v0Handler
-    :<|> v1Handler
+    v1Handler
     :<|> internalHandler
   where
-    -- TODO: It'd be nicer to not throw an exception here, but servant doesn't
-    -- make this very easy at the moment.
-    -- We can use https://hackage.haskell.org/package/servant-generate to address this.
-    v0Handler       = error "V0 API no longer supported"
     v1Handler       = V1.handlers w
     internalHandler = Internal.handlers (walletPassiveLayer w)
 
-walletDocServer :: (HasCompileInfo, HasUpdateConfiguration) => Server WalletDocAPI
-walletDocServer = v0DocHandler :<|> v1DocHandler
+walletDocServer :: (HasCompileInfo, HasUpdateConfiguration) => Server NewWalletDocAPI
+walletDocServer =
+    v1DocHandler
   where
-    v0DocHandler = error "V0 API no longer supported"
-    v1DocHandler = swaggerSchemaUIServer (Swagger.api (compileInfo, curSoftwareVersion) v1API Swagger.highLevelDescription)
+    infos        = (compileInfo, curSoftwareVersion)
+    v1DocHandler = swaggerSchemaUIServer
+        (Swagger.api infos newWalletAPI Swagger.highLevelDescription)
