+++ wallet-new/src/Cardano/Wallet/Server/Plugins.hs
@@ -2,267 +2,134 @@
      A @Plugin@ is essentially a set of actions which will be run in
      a particular monad, at some point in time.
 -}
-{-# LANGUAGE TupleSections #-}
 
-module Cardano.Wallet.Server.Plugins (
-      Plugin
-    , syncWalletWorker
-    , acidCleanupWorker
-    , conversation
-    , legacyWalletBackend
-    , walletBackend
-    , walletDocumentation
-    , resubmitterPlugin
-    , notifierPlugin
+module Cardano.Wallet.Server.Plugins
+    ( Plugin
+    , apiServer
+    , docServer
+    , acidStateSnapshots
+    , updateNotifier
     ) where
 
 import           Universum
 
-import           Cardano.Wallet.API as API
-import           Cardano.Wallet.API.V1.Headers (applicationJson)
-import qualified Cardano.Wallet.API.V1.Types as V1
-import           Cardano.Wallet.Kernel (PassiveWallet)
-import qualified Cardano.Wallet.Kernel.Diffusion as Kernel
-import qualified Cardano.Wallet.Kernel.Mode as Kernel
-import qualified Cardano.Wallet.LegacyServer as LegacyServer
-import qualified Cardano.Wallet.Server as Server
-import           Cardano.Wallet.Server.CLI (NewWalletBackendParams (..),
-                     RunMode, WalletBackendParams (..), isDebugMode,
-                     walletAcidInterval, walletDbOptions)
-import           Cardano.Wallet.WalletLayer (ActiveWalletLayer,
-                     PassiveWalletLayer)
-import qualified Cardano.Wallet.WalletLayer.Kernel as WalletLayer.Kernel
-import qualified Pos.Wallet.Web.Error.Types as V0
-
-import           Control.Exception (fromException)
-import           Data.Aeson
-import           Formatting (build, sformat, (%))
-import           Network.HTTP.Types.Status (badRequest400)
-import           Network.Wai (Application, Middleware, Response, responseLBS)
-import           Network.Wai.Handler.Warp (defaultSettings,
-                     setOnExceptionResponse)
+import           Network.Wai (Application, Middleware)
+import           Network.Wai.Handler.Warp (defaultSettings)
 import           Network.Wai.Middleware.Cors (cors, corsMethods,
                      corsRequestHeaders, simpleCorsResourcePolicy,
                      simpleMethods)
-import           Ntp.Client (NtpStatus)
-import           Pos.Chain.Txp (TxpConfiguration)
-import           Pos.Infra.Diffusion.Types (Diffusion (..))
-import           Pos.Wallet.Web (cleanupAcidStatePeriodically)
-import           Pos.Wallet.Web.Pending.Worker (startPendingTxsResubmitter)
-import qualified Pos.Wallet.Web.Server.Runner as V0
-import           Pos.Wallet.Web.Sockets (getWalletWebSockets,
-                     upgradeApplicationWS)
-import qualified Servant
-import           System.Wlog (logInfo, modifyLoggerName, usingLoggerName)
-
-import           Pos.Context (HasNodeContext)
-import           Pos.Crypto (ProtocolMagic)
-import           Pos.Util (lensOf)
+import           System.Wlog (logInfo, usingLoggerName)
 
 import           Cardano.NodeIPC (startNodeJsIPC)
-import           Pos.Configuration (walletProductionApi,
-                     walletTxCreationDisabled)
-import           Pos.Infra.Shutdown.Class (HasShutdownContext (shutdownContext))
+import           Cardano.Wallet.API as API
+import           Cardano.Wallet.Kernel (PassiveWallet)
+import           Cardano.Wallet.Server.CLI (NewWalletBackendParams (..),
+                     RunMode, WalletBackendParams (..), isDebugMode)
+import           Cardano.Wallet.WalletLayer (ActiveWalletLayer,
+                     PassiveWalletLayer)
+import           Pos.Crypto (ProtocolMagic)
+import           Pos.Infra.Diffusion.Types (Diffusion (..))
+import           Pos.Infra.Shutdown (HasShutdownContext (shutdownContext),
+                     ShutdownContext)
 import           Pos.Launcher.Configuration (HasConfigurations)
 import           Pos.Util.CompileInfo (HasCompileInfo)
-import           Pos.Wallet.Web.Mode (WalletWebMode)
-import           Pos.Wallet.Web.Server.Launcher (walletDocumentationImpl,
-                     walletServeImpl)
-import           Pos.Wallet.Web.State (askWalletDB)
-import           Pos.Wallet.Web.Tracking.Sync (processSyncRequest)
-import           Pos.Wallet.Web.Tracking.Types (SyncQueue)
-import           Pos.Web (serveWeb)
-import           Pos.WorkMode (WorkMode)
+import           Pos.Web (serveDocImpl, serveImpl)
+
+import qualified Cardano.Wallet.Kernel.Diffusion as Kernel
+import qualified Cardano.Wallet.Kernel.Mode as Kernel
+import qualified Cardano.Wallet.Server as Server
+import qualified Cardano.Wallet.WalletLayer.Kernel as WalletLayer.Kernel
+import qualified Data.ByteString.Char8 as BS8
+import qualified Servant
+
+-- Needed for Orphan Instance 'Buildable Servant.NoContent' :|
+import           Pos.Wallet.Web ()
 
 
 -- A @Plugin@ running in the monad @m@.
 type Plugin m = [Diffusion m -> m ()]
 
--- | A @Plugin@ to periodically compact & snapshot the acid-state database.
-acidCleanupWorker :: WalletBackendParams
-                  -> Plugin WalletWebMode
-acidCleanupWorker WalletBackendParams{..} = pure $ const $
-    modifyLoggerName (const "acidcleanup") $
-    askWalletDB >>= \db -> cleanupAcidStatePeriodically db (walletAcidInterval walletDbOptions)
 
--- | The @Plugin@ which defines part of the conversation protocol for this node.
-conversation :: HasConfigurations => WalletBackendParams -> Plugin WalletWebMode
-conversation wArgs = map const (pluginsMonitoringApi wArgs)
-  where
-    pluginsMonitoringApi :: (WorkMode ctx m , HasNodeContext ctx)
-                         => WalletBackendParams
-                         -> [m ()]
-    pluginsMonitoringApi WalletBackendParams {..}
-        | enableMonitoringApi = [serveWeb monitoringApiPort walletTLSParams]
-        | otherwise = []
-
-walletDocumentation
-    :: (HasConfigurations, HasCompileInfo)
-    => WalletBackendParams
-    -> Plugin WalletWebMode
-walletDocumentation WalletBackendParams {..} = pure $ \_ ->
-    walletDocumentationImpl
-        application
-        walletDocAddress
-        tls
-        (Just defaultSettings)
-        Nothing
-  where
-    application :: WalletWebMode Application
-    application = do
-        let app = Servant.serve API.walletDocAPI LegacyServer.walletDocServer
-        return $ withMiddleware walletRunMode app
-    tls =
-        if isDebugMode walletRunMode then Nothing else walletTLSParams
-
--- | A @Plugin@ to start the wallet backend API.
-legacyWalletBackend :: (HasConfigurations, HasCompileInfo)
-                    => ProtocolMagic
-                    -> TxpConfiguration
-                    -> WalletBackendParams
-                    -> TVar NtpStatus
-                    -> Plugin WalletWebMode
-legacyWalletBackend pm txpConfig WalletBackendParams {..} ntpStatus = pure $ \diffusion -> do
-    modifyLoggerName (const "legacyServantBackend") $ do
-      logInfo $ sformat ("Production mode for API: "%build)
-        walletProductionApi
-      logInfo $ sformat ("Transaction submission disabled: "%build)
-        walletTxCreationDisabled
-
-      ctx <- view shutdownContext
-      let
-        portCallback :: Word16 -> IO ()
-        portCallback port = usingLoggerName "NodeIPC" $ flip runReaderT ctx $ startNodeJsIPC port
-      walletServeImpl
-        (getApplication diffusion)
-        walletAddress
-        -- Disable TLS if in debug mode.
-        (if isDebugMode walletRunMode then Nothing else walletTLSParams)
-        (Just $ setOnExceptionResponse exceptionHandler defaultSettings)
-        (Just portCallback)
-  where
-    -- Gets the Wai `Application` to run.
-    getApplication :: Diffusion WalletWebMode -> WalletWebMode Application
-    getApplication diffusion = do
-        logInfo "Wallet Web API has STARTED!"
-        wsConn <- getWalletWebSockets
-        ctx <- V0.walletWebModeContext
-        return
-            $ withMiddleware walletRunMode
-            $ upgradeApplicationWS wsConn
-            $ Servant.serve API.walletAPI
-            $ LegacyServer.walletServer
-                (V0.convertHandler ctx)
-                pm
-                txpConfig
-                diffusion
-                ntpStatus
-                walletRunMode
-
-    exceptionHandler :: SomeException -> Response
-    exceptionHandler se =
-        case asum [handleV1Errors se, handleV0Errors se] of
-             Nothing -> handleGenericError se
-             Just r  -> r
-
-    -- Handles domain-specific errors coming from the V1 API.
-    handleV1Errors :: SomeException -> Maybe Response
-    handleV1Errors se =
-        let reify (we :: V1.WalletError) =
-                responseLBS (V1.toHttpErrorStatus we) [applicationJson] .  encode $ we
-        in fmap reify (fromException se)
-
-    -- Handles domain-specific errors coming from the V0 API, but rewraps it
-    -- into a jsend payload. It doesn't explicitly handle 'InternalError' or
-    -- 'DecodeError', as they can come from any part of the stack or even
-    -- rewrap some other exceptions (cfr 'rewrapToWalletError').
-    -- Uses the 'Buildable' istance on 'WalletError' to exploit any
-    -- available rendering and information-masking improvements.
-    handleV0Errors :: SomeException -> Maybe Response
-    handleV0Errors se =
-        let maskSensitive err =
-                case err of
-                    V0.RequestError _         -> err
-                    V0.DuplicateWalletError _ -> V0.RequestError "DuplicateWalletError"
-                    V0.NoSuchWalletError _    -> V0.RequestError "NoSuchWalletError"
-                    V0.InternalError _        -> V0.RequestError "InternalError"
-                    V0.DecodeError _          -> V0.RequestError "DecodeError"
-            reify :: V0.WalletError -> V1.WalletError
-            reify = V1.UnknownError . sformat build . maskSensitive
-        in fmap (responseLBS badRequest400 [applicationJson] .  encode . reify) (fromException se)
-
-    -- Handles any generic error, trying to prevent internal exceptions from leak outside.
-    handleGenericError :: SomeException -> Response
-    handleGenericError _ =
-        let
-            unknownV1Error = V1.UnknownError "Something went wrong."
-        in
-            responseLBS badRequest400 [applicationJson] $ encode unknownV1Error
-
-
--- | A 'Plugin' to start the wallet REST server
---
--- NOTE: There is no web socket support in the new wallet for now.
-walletBackend :: ProtocolMagic
-              -> NewWalletBackendParams
-              -> (PassiveWalletLayer IO, PassiveWallet)
-              -> Plugin Kernel.WalletMode
-walletBackend protocolMagic (NewWalletBackendParams WalletBackendParams{..}) (passiveLayer, passiveWallet) =
+-- | A @Plugin@ to start the wallet REST server
+apiServer
+    :: ProtocolMagic
+    -> NewWalletBackendParams
+    -> (PassiveWalletLayer IO, PassiveWallet)
+    -> Plugin Kernel.WalletMode
+apiServer protocolMagic (NewWalletBackendParams WalletBackendParams{..}) (passiveLayer, passiveWallet) =
     pure $ \diffusion -> do
         env <- ask
         let diffusion' = Kernel.fromDiffusion (lower env) diffusion
         WalletLayer.Kernel.bracketActiveWallet protocolMagic passiveLayer passiveWallet diffusion' $ \active _ -> do
           ctx <- view shutdownContext
-          let
-            portCallback :: Word16 -> IO ()
-            portCallback port = usingLoggerName "NodeIPC" $ flip runReaderT ctx $ startNodeJsIPC port
-          walletServeImpl
+          serveImpl
             (getApplication active)
-            walletAddress
-            -- Disable TLS if in debug modeit .
+            (BS8.unpack ip)
+            port
             (if isDebugMode walletRunMode then Nothing else walletTLSParams)
             Nothing
-            (Just portCallback)
+            (Just $ portCallback ctx)
   where
+    (ip, port) = walletDocAddress
+
     getApplication :: ActiveWalletLayer IO -> Kernel.WalletMode Application
     getApplication active = do
       logInfo "New wallet API has STARTED!"
       return $ withMiddleware walletRunMode $
-        Servant.serve API.walletAPI $ Server.walletServer active walletRunMode
+        Servant.serve API.newWalletAPI $ Server.walletServer active walletRunMode
 
     lower :: env -> ReaderT env IO a -> IO a
     lower env m = runReaderT m env
 
--- | A @Plugin@ to resubmit pending transactions.
-resubmitterPlugin :: HasConfigurations
-                  => ProtocolMagic
-                  -> TxpConfiguration
-                  -> Plugin WalletWebMode
-resubmitterPlugin pm txpConfig = [\diffusion -> askWalletDB >>= \db ->
-                        startPendingTxsResubmitter pm txpConfig db (sendTx diffusion)]
+    portCallback :: ShutdownContext -> Word16 -> IO ()
+    portCallback ctx =
+        usingLoggerName "NodeIPC" . flip runReaderT ctx . startNodeJsIPC
+
+    -- | "Attaches" the middleware to this 'Application', if any.
+    -- When running in debug mode, chances are we want to at least allow CORS to test the API
+    -- with a Swagger editor, locally.
+    withMiddleware :: RunMode -> Application -> Application
+    withMiddleware wrm app
+      | isDebugMode wrm = corsMiddleware app
+      | otherwise = app
+
+    corsMiddleware :: Middleware
+    corsMiddleware = cors (const $ Just policy)
+        where
+          policy = simpleCorsResourcePolicy
+            { corsRequestHeaders = ["Content-Type"]
+            , corsMethods = "PUT" : simpleMethods
+            }
+
+
+-- | A @Plugin@ to serve the wallet documentation
+docServer
+    :: (HasConfigurations, HasCompileInfo)
+    => NewWalletBackendParams
+    -> Plugin Kernel.WalletMode
+docServer (NewWalletBackendParams WalletBackendParams{..}) = pure $ \_ ->
+    serveDocImpl
+        application
+        (BS8.unpack ip)
+        port
+        (if isDebugMode walletRunMode then Nothing else walletTLSParams)
+        (Just defaultSettings)
+        Nothing
+  where
+    (ip, port) = walletDocAddress
+
+    application :: Kernel.WalletMode Application
+    application =
+        return $ Servant.serve API.newWalletDocAPI Server.walletDocServer
+
+
+-- | A @Plugin@ to periodically compact & snapshot the acid-state database.
+acidStateSnapshots :: Plugin Kernel.WalletMode
+acidStateSnapshots =
+    error "Not Implemented: acidStateSnapshots [CBR-305]"
+
 
 -- | A @Plugin@ to notify frontend via websockets.
-notifierPlugin :: HasConfigurations => Plugin WalletWebMode
-notifierPlugin = [const V0.notifierPlugin]
-
--- | The @Plugin@ responsible for the restoration & syncing of a wallet.
-syncWalletWorker :: HasConfigurations => Plugin WalletWebMode
-syncWalletWorker = pure $ const $
-    modifyLoggerName (const "syncWalletWorker") $
-    (view (lensOf @SyncQueue) >>= processSyncRequest)
-
--- | "Attaches" the middleware to this 'Application', if any.
--- When running in debug mode, chances are we want to at least allow CORS to test the API
--- with a Swagger editor, locally.
-withMiddleware :: RunMode -> Application -> Application
-withMiddleware wrm app
-  | isDebugMode wrm = corsMiddleware app
-  | otherwise = app
-
-corsMiddleware :: Middleware
-corsMiddleware = cors (const $ Just policy)
-    where
-      policy = simpleCorsResourcePolicy
-        { corsRequestHeaders = ["Content-Type"]
-        , corsMethods = "PUT" : simpleMethods
-        }
+updateNotifier :: Plugin Kernel.WalletMode
+updateNotifier =
+    error "Not Implemented: updateNotifier [CBR-374]"
